requirements.txt
# Dependencies for Fundamental Analysis Engine
requests>=2.31.0
nats-py>=2.6.0
orjson>=3.10.7
PyQt5>=5.15.0
numpy>=1.24.0
scipy>=1.10.0
psutil>=5.9.0

